INCOMPLETE.  DOES NOT COMPILE.  Starting to restructure and play around with KIF.
Restructuring KIF to be more object-oriented and to give more options
to users for implementing steps and scenarios.  Part of the goal of
this is to keep the original KIF interface intact keeping existing code
compatible, even if the underlying method of implementation will be
different.

This commit represents incomplete work and does not compile.  It is not
recommended to pull this commit.